Add container view and button view padding adjustment

+++ src/ComponentSystem/Component.cs
@@ -11,6 +11,6 @@ namespace Industropolis.Engine
             get => _enabled && Parent != null ? Parent.Enabled : _enabled;
             set => _enabled = value;
         }
-        public float Priority { get; set; }
+        public double Priority { get; set; }
     }
 }
