+++ src/Node.cs
@@ -16,7 +16,7 @@ namespace Industropolis.Engine
         public virtual Vector2 Size { get; set; } = Vector2.Zero;
 
         public int Layer { get; set; } = -1;
-        public float SceneSort { get; set; }
+        public double SceneSort { get; set; }
 
         public int Depth => Parent != null ? Parent.Depth + 1 : 0;
